notices/rest_api/v1/serializers.py
"""Serializers for the notices v1 rest API"""
from rest_framework import serializers

from notices.models import Notice


class NoticeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notice
        fields = ('id', 'name', 'html_content')
