requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    make upgrade
#
attrs==21.2.0
    # via pytest
bleach==4.1.0
    # via -r requirements/base.txt
certifi==2021.5.30
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.14.6
    # via
    #   -r requirements/base.txt
    #   cryptography
charset-normalizer==2.0.4
    # via
    #   -r requirements/base.txt
    #   requests
click==8.0.1
    # via code-annotations
code-annotations==1.2.0
    # via -r requirements/test.in
coverage==5.5
    # via pytest-cov
cryptography==3.4.8
    # via
    #   -r requirements/base.txt
    #   pyjwt
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   rest-condition
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-model-utils==4.1.1
    # via -r requirements/base.txt
django-rest-framework==0.1.0
    # via -r requirements/base.txt
django-simple-history==3.0.0
    # via -r requirements/base.txt
django-waffle==2.2.1
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.12.4
    # via
    #   -r requirements/base.txt
    #   django-rest-framework
    #   drf-jwt
    #   edx-drf-extensions
    #   rest-condition
drf-jwt==1.19.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-django-utils==4.3.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-drf-extensions==6.6.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
edx-opaque-keys==2.2.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
future==0.18.2
    # via
    #   -r requirements/base.txt
    #   pyjwkest
idna==3.2
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==1.1.1
    # via pytest
jinja2==3.0.1
    # via code-annotations
markupsafe==2.0.1
    # via jinja2
newrelic==6.8.1.164
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
packaging==21.0
    # via
    #   -r requirements/base.txt
    #   bleach
    #   pytest
pbr==5.6.0
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==0.13.1
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   pytest
psutil==5.8.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
py==1.10.0
    # via pytest
pycparser==2.20
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.10.1
    # via
    #   -r requirements/base.txt
    #   pyjwkest
pyjwkest==1.4.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
pyjwt[crypto]==1.7.1
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   drf-jwt
pymongo==3.12.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pyparsing==2.4.7
    # via
    #   -r requirements/base.txt
    #   packaging
pytest==6.2.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==2.12.1
    # via -r requirements/test.in
pytest-django==4.4.0
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
python-slugify==5.0.2
    # via code-annotations
pytz==2021.1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==5.4.1
    # via code-annotations
requests==2.26.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   pyjwkest
rest-condition==1.0.3
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
semantic-version==2.8.5
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   bleach
    #   edx-drf-extensions
    #   pyjwkest
    #   python-dateutil
sqlparse==0.4.1
    # via
    #   -r requirements/base.txt
    #   django
stevedore==3.4.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
urllib3==1.26.6
    # via
    #   -r requirements/base.txt
    #   requests
webencodings==0.5.1
    # via
    #   -r requirements/base.txt
    #   bleach
